pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lot"
version = "0.1.0"
description = "MVP scaffold for an agent-oriented embedded virtual debugging platform."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "pydantic>=2.7,<3.0",
    "PyYAML>=6.0,<7.0",
    "uvicorn>=0.30,<1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
